Dublicate code was removed from Project and Matrix Build. BaseBuildableProject was introduced.
