+++ hudson-core/src/main/java/hudson/matrix/MatrixBuild.java
@@ -26,6 +26,7 @@ package hudson.matrix;
 import hudson.Util;
 import hudson.model.AbstractBuild;
 import hudson.model.AbstractProject;
+import hudson.model.Build;
 import hudson.model.BuildListener;
 import hudson.model.Executor;
 import hudson.model.Fingerprint;
@@ -57,7 +58,7 @@ import org.kohsuke.stapler.StaplerResponse;
  *
  * @author Kohsuke Kawaguchi
  */
-public class MatrixBuild extends AbstractBuild<MatrixProject,MatrixBuild> {
+public class MatrixBuild extends Build<MatrixProject,MatrixBuild> {
     private AxisList axes;
 
     public MatrixBuild(MatrixProject job) throws IOException {
