+++ hudson-core/src/main/java/hudson/matrix/MatrixProject.java
@@ -28,10 +28,8 @@ import hudson.CopyOnWrite;
 import hudson.Extension;
 import hudson.Util;
 import hudson.XmlFile;
-import hudson.model.AbstractProject;
-import hudson.model.Action;
+import hudson.model.BaseBuildableProject;
 import hudson.model.BuildableItemWithBuildWrappers;
-import hudson.model.DependencyGraph;
 import hudson.model.Descriptor;
 import hudson.model.Descriptor.FormException;
 import hudson.model.Hudson;
@@ -46,13 +44,10 @@ import hudson.model.Result;
 import hudson.model.SCMedItem;
 import hudson.model.Saveable;
 import hudson.model.TopLevelItem;
-import hudson.tasks.BuildStep;
 import hudson.tasks.BuildStepDescriptor;
-import hudson.tasks.BuildWrapper;
 import hudson.tasks.BuildWrappers;
 import hudson.tasks.Builder;
 import hudson.tasks.Publisher;
-import hudson.triggers.Trigger;
 import hudson.util.CopyOnWriteMap;
 import hudson.util.DescribableList;
 import hudson.util.DescribableListUtil;
@@ -88,7 +83,7 @@ import org.springframework.util.CollectionUtils;
  *
  * @author Kohsuke Kawaguchi
  */
-public class MatrixProject extends AbstractProject<MatrixProject, MatrixBuild> implements IMatrixProject, TopLevelItem,
+public class MatrixProject extends BaseBuildableProject<MatrixProject, MatrixBuild> implements IMatrixProject, TopLevelItem,
     SCMedItem, ItemGroup<MatrixConfiguration>, Saveable, FlyweightTask, BuildableItemWithBuildWrappers {
 
     public static final String HAS_COMBINATION_FILTER_PARAM = "hasCombinationFilter";
@@ -120,39 +115,6 @@ public class MatrixProject extends AbstractProject<MatrixProject, MatrixBuild> i
      */
     @Deprecated
     private volatile String combinationFilter;
-
-    /**
-     * List of active {@link Builder}s configured for this project.
-     *
-     * @deprecated as of 2.2.0
-     *             don't use this field directly, logic was moved to {@link org.hudsonci.api.model.IProjectProperty}.
-     *             Use getter/setter for accessing to this field.
-     */
-    @Deprecated
-    private DescribableList<Builder,Descriptor<Builder>> builders =
-            new DescribableList<Builder,Descriptor<Builder>>(this);
-
-    /**
-     * List of active {@link Publisher}s configured for this project.
-     *
-     * @deprecated as of 2.2.0
-     *             don't use this field directly, logic was moved to {@link org.hudsonci.api.model.IProjectProperty}.
-     *             Use getter/setter for accessing to this field.
-     */
-    @Deprecated
-    private DescribableList<Publisher,Descriptor<Publisher>> publishers =
-            new DescribableList<Publisher,Descriptor<Publisher>>(this);
-
-    /**
-     * List of active {@link BuildWrapper}s configured for this project.
-     *
-     * @deprecated as of 2.2.0
-     *             don't use this field directly, logic was moved to {@link org.hudsonci.api.model.IProjectProperty}.
-     *             Use getter/setter for accessing to this field.
-     */
-    @Deprecated
-    private DescribableList<BuildWrapper,Descriptor<BuildWrapper>> buildWrappers =
-            new DescribableList<BuildWrapper,Descriptor<BuildWrapper>>(this);
 
     /**
      * All {@link MatrixConfiguration}s, keyed by their {@link MatrixConfiguration#getName() names}.
@@ -291,62 +253,6 @@ public class MatrixProject extends AbstractProject<MatrixProject, MatrixBuild> i
         getStringProperty(CUSTOM_WORKSPACE_PROPERTY_NAME).setValue(customWorkspace);
     }
 
-    /**
-     * @inheritDoc
-     */
-    public List<Builder> getBuilders() {
-        return getBuildersList().toList();
-    }
-
-	@SuppressWarnings("unchecked")
-    public DescribableList<Builder,Descriptor<Builder>> getBuildersList() {
-        return getDescribableListProjectProperty(BUILDERS_PROPERTY_NAME).getValue();
-    }
-
-
-    public void setBuilders(DescribableList<Builder,Descriptor<Builder>> builders) {
-        getDescribableListProjectProperty(BUILDERS_PROPERTY_NAME).setValue(builders);
-    }
-
-    /**
-     * @inheritDoc
-     */
-    public Map<Descriptor<Publisher>,Publisher> getPublishers() {
-        return getPublishersList().toMap();
-    }
-
-    /**
-     * @inheritDoc
-     */
-	@SuppressWarnings("unchecked")
-    public DescribableList<Publisher, Descriptor<Publisher>> getPublishersList() {
-        return getDescribableListProjectProperty(PUBLISHERS_PROPERTY_NAME).getValue();
-    }
-
-    public void setPublishers(DescribableList<Publisher, Descriptor<Publisher>> publishers) {
-        getDescribableListProjectProperty(PUBLISHERS_PROPERTY_NAME).setValue(publishers);
-    }
-
-    /**
-     * @inheritDoc
-     */
-	@SuppressWarnings("unchecked")
-    public DescribableList<BuildWrapper, Descriptor<BuildWrapper>> getBuildWrappersList() {
-        return getDescribableListProjectProperty(BUILD_WRAPPERS_PROPERTY_NAME).getValue();
-    }
-
-    /**
-     * @inheritDoc
-     */
-    public Map<Descriptor<BuildWrapper>,BuildWrapper> getBuildWrappers() {
-        return getBuildWrappersList().toMap();
-    }
-
-    public void setBuildWrappers(DescribableList<BuildWrapper, Descriptor<BuildWrapper>> buildWrappers) {
-        getDescribableListProjectProperty(BUILD_WRAPPERS_PROPERTY_NAME).setValue(buildWrappers);
-    }
-
-
     @Override
     protected void buildProjectProperties() throws IOException {
         super.buildProjectProperties();
@@ -375,36 +281,8 @@ public class MatrixProject extends AbstractProject<MatrixProject, MatrixBuild> i
             setCustomWorkspace(customWorkspace);
             customWorkspace = null;//Reset to null. No longer needed.
         }
-        if (null == getProperty(BUILDERS_PROPERTY_NAME)) {
-            setBuilders(builders);
-            builders = null;
-        }
-        if (null == getProperty(BUILD_WRAPPERS_PROPERTY_NAME)) {
-            setBuildWrappers(buildWrappers);
-            buildWrappers = null;
-        }
-        if (null == getProperty(PUBLISHERS_PROPERTY_NAME)) {
-            setPublishers(publishers);
-            publishers = null;
-        }
         save();
         rebuildConfigurations();
-    }
-
-    @Override
-    protected List<Action> createTransientActions() {
-        List<Action> r = super.createTransientActions();
-
-        for (BuildStep step : getBuildersList())
-            r.addAll(step.getProjectActions(this));
-        for (BuildStep step : getPublishersList())
-            r.addAll(step.getProjectActions(this));
-        for (BuildWrapper step : getBuildWrappersList())
-            r.addAll(step.getProjectActions(this));
-        for (Trigger trigger : getTriggersList())
-            r.addAll(trigger.getProjectActions());
-
-        return r;
     }
 
     /**
@@ -437,9 +315,6 @@ public class MatrixProject extends AbstractProject<MatrixProject, MatrixBuild> i
             // perhaps the file was edited on disk and the sort order might have been broken
             Collections.sort(getAxes());
         }
-        getBuildersList().setOwner(this);
-        getPublishersList().setOwner(this);
-        getBuildWrappersList().setOwner(this);
         rebuildConfigurations();
     }
 
@@ -659,7 +534,7 @@ public class MatrixProject extends AbstractProject<MatrixProject, MatrixBuild> i
     }
 
     public Publisher getPublisher(Descriptor<Publisher> descriptor) {
-        for (Publisher p : publishers) {
+        for (Publisher p : getPublishersList()) {
             if(p.getDescriptor()==descriptor)
                 return p;
         }
@@ -672,12 +547,6 @@ public class MatrixProject extends AbstractProject<MatrixProject, MatrixBuild> i
 
     public boolean isFingerprintConfigured() {
         return false;
-    }
-
-    protected void buildDependencyGraph(DependencyGraph graph) {
-        getPublishersList().buildDependencyGraph(this,graph);
-        getBuildersList().buildDependencyGraph(this,graph);
-        getBuildWrappersList().buildDependencyGraph(this,graph);
     }
 
     public MatrixProject asProject() {
@@ -726,8 +595,7 @@ public class MatrixProject extends AbstractProject<MatrixProject, MatrixBuild> i
 
         setBuildWrappers(DescribableListUtil.buildFromJson(this, req, json, BuildWrappers.getFor(this)));
         setBuilders(DescribableListUtil.buildFromHetero(this, req, json, "builder", Builder.all()));
-        setPublishers(DescribableListUtil.buildFromJson(this, req, json,
-            BuildStepDescriptor.filter(Publisher.all(), this.getClass())));
+        buildPublishers(req, json, BuildStepDescriptor.filter(Publisher.all(), this.getClass()));
 
         rebuildConfigurations();
     }
