+++ hudson-core/src/main/java/hudson/model/Project.java
@@ -24,13 +24,11 @@
  */
 package hudson.model;
 
-import hudson.Functions;
 import hudson.Util;
 import hudson.diagnosis.OldDataMonitor;
 import hudson.model.Descriptor.FormException;
 import hudson.tasks.BuildStep;
 import hudson.tasks.BuildStepDescriptor;
-import hudson.tasks.BuildWrapper;
 import hudson.tasks.BuildWrappers;
 import hudson.tasks.Builder;
 import hudson.tasks.Fingerprinter;
@@ -38,19 +36,13 @@ import hudson.tasks.Maven;
 import hudson.tasks.Maven.MavenInstallation;
 import hudson.tasks.Maven.ProjectWithMaven;
 import hudson.tasks.Publisher;
-import hudson.triggers.Trigger;
-import hudson.util.DescribableList;
 import hudson.util.DescribableListUtil;
 import java.io.IOException;
 import java.util.HashSet;
-import java.util.List;
-import java.util.Map;
 import java.util.Set;
-import java.util.concurrent.CopyOnWriteArrayList;
 import javax.servlet.ServletException;
 import net.sf.json.JSONObject;
 import org.hudsonci.api.model.IProject;
-import org.hudsonci.model.project.property.BaseProjectProperty;
 import org.kohsuke.stapler.StaplerRequest;
 import org.kohsuke.stapler.StaplerResponse;
 
@@ -60,41 +52,8 @@ import org.kohsuke.stapler.StaplerResponse;
  * @author Kohsuke Kawaguchi
  */
 public abstract class Project<P extends Project<P,B>,B extends Build<P,B>>
-    extends AbstractProject<P, B>
+    extends BaseBuildableProject<P, B>
     implements SCMedItem, Saveable, ProjectWithMaven, BuildableItemWithBuildWrappers, IProject {
-
-    /**
-     * List of active {@link Builder}s configured for this project.
-     *
-     * @deprecated as of 2.2.0
-     *             don't use this field directly, logic was moved to {@link org.hudsonci.api.model.IProjectProperty}.
-     *             Use getter/setter for accessing to this field.
-     */
-    @Deprecated
-    private DescribableList<Builder, Descriptor<Builder>> builders =
-        new DescribableList<Builder, Descriptor<Builder>>(this);
-
-    /**
-     * List of active {@link Publisher}s configured for this project.
-     *
-     * @deprecated as of 2.2.0
-     *             don't use this field directly, logic was moved to {@link org.hudsonci.api.model.IProjectProperty}.
-     *             Use getter/setter for accessing to this field.
-     */
-    @Deprecated
-    private DescribableList<Publisher, Descriptor<Publisher>> publishers =
-        new DescribableList<Publisher, Descriptor<Publisher>>(this);
-
-    /**
-     * List of active {@link BuildWrapper}s configured for this project.
-     *
-     * @deprecated as of 2.2.0
-     *             don't use this field directly, logic was moved to {@link org.hudsonci.api.model.IProjectProperty}.
-     *             Use getter/setter for accessing to this field.
-     */
-    @Deprecated
-    private DescribableList<BuildWrapper, Descriptor<BuildWrapper>> buildWrappers =
-        new DescribableList<BuildWrapper, Descriptor<BuildWrapper>>(this);
 
     /**
      * Creates a new project.
@@ -103,89 +62,8 @@ public abstract class Project<P extends Project<P,B>,B extends Build<P,B>>
         super(parent, name);
     }
 
-    @Override
-    public void onLoad(ItemGroup<? extends Item> parent, String name) throws IOException {
-        super.onLoad(parent, name);
-
-        getBuildersList().setOwner(this);
-        getPublishersList().setOwner(this);
-        getBuildWrappersList().setOwner(this);
-    }
-
-    @Override
-    protected void buildProjectProperties() throws IOException {
-        super.buildProjectProperties();
-        convertBuildersProjectProperty();
-        convertBuildWrappersProjectProperties();
-        convertPublishersProperties();
-    }
-
-    void convertPublishersProperties() {
-        if (null != publishers) {
-            putAllProjectProperties(DescribableListUtil.convertToProjectProperties(publishers, this), false);
-            publishers = null;
-        }
-    }
-
-    void convertBuildWrappersProjectProperties() {
-        if (null == getProperty(BUILD_WRAPPERS_PROPERTY_NAME)) {
-            setBuildWrappers(buildWrappers);
-            buildWrappers = null;
-        }
-    }
-
-    void convertBuildersProjectProperty() {
-        if (null == getProperty(BUILDERS_PROPERTY_NAME)) {
-            setBuilders(builders);
-            builders = null;
-        }
-    }
-
     public AbstractProject<?, ?> asProject() {
         return this;
-    }
-
-    public List<Builder> getBuilders() {
-        return getBuildersList().toList();
-    }
-
-    public Map<Descriptor<Publisher>,Publisher> getPublishers() {
-        return getPublishersList().toMap();
-    }
-
-    public void setBuilders(DescribableList<Builder,Descriptor<Builder>> builders) {
-        getDescribableListProjectProperty(BUILDERS_PROPERTY_NAME).setValue(builders);
-    }
-
-    public DescribableList<Builder,Descriptor<Builder>> getBuildersList() {
-        return getDescribableListProjectProperty(BUILDERS_PROPERTY_NAME).getValue();
-    }
-
-    public DescribableList<Publisher, Descriptor<Publisher>> getPublishersList() {
-        List<Descriptor<Publisher>> descriptors = Functions.getPublisherDescriptors(this);
-        List<Publisher> publisherList = new CopyOnWriteArrayList<Publisher>();
-        DescribableList<Publisher, Descriptor<Publisher>> result
-            = new DescribableList<Publisher, Descriptor<Publisher>>(this);
-        for (Descriptor<Publisher> descriptor : descriptors) {
-            BaseProjectProperty<Publisher> property = getBaseProjectProperty(descriptor.getJsonSafeClassName());
-            if (null != property.getValue()) {
-                publisherList.add(property.getValue());
-            }
-        }
-        result.addAllTo(publisherList);
-        return result;
-    }
-
-    public Map<Descriptor<BuildWrapper>,BuildWrapper> getBuildWrappers() {
-        return getBuildWrappersList().toMap();
-    }
-
-    public DescribableList<BuildWrapper, Descriptor<BuildWrapper>> getBuildWrappersList() {
-        return getDescribableListProjectProperty(BUILD_WRAPPERS_PROPERTY_NAME).getValue();
-    }
-
-    public void setBuildWrappers(DescribableList<BuildWrapper, Descriptor<BuildWrapper>> buildWrappers) {
-        getDescribableListProjectProperty(BUILD_WRAPPERS_PROPERTY_NAME).setValue(buildWrappers);
     }
 
     @Override
@@ -220,15 +98,6 @@ public abstract class Project<P extends Project<P,B>,B extends Build<P,B>>
         getPublishersList().remove(descriptor);
     }
 
-    public Publisher getPublisher(Descriptor<Publisher> descriptor) {
-        return (Publisher) getBaseProjectProperty(descriptor.getJsonSafeClassName()).getValue();
-    }
-
-    protected void buildDependencyGraph(DependencyGraph graph) {
-        getPublishersList().buildDependencyGraph(this, graph);
-        getBuildersList().buildDependencyGraph(this, graph);
-        getBuildWrappersList().buildDependencyGraph(this, graph);
-    }
 
     @Override
     public boolean isFingerprintConfigured() {
@@ -241,11 +110,6 @@ public abstract class Project<P extends Project<P,B>,B extends Build<P,B>>
         return null;
     }
 
-//
-//
-// actions
-//
-//
     @Override
     protected void submit( StaplerRequest req, StaplerResponse rsp ) throws IOException, ServletException, FormException {
         super.submit(req,rsp);
@@ -253,34 +117,6 @@ public abstract class Project<P extends Project<P,B>,B extends Build<P,B>>
         setBuildWrappers(DescribableListUtil.buildFromJson(this, req, json, BuildWrappers.getFor(this)));
         setBuilders(DescribableListUtil.buildFromHetero(this, req, json, "builder", Builder.all()));
         buildPublishers(req, json, BuildStepDescriptor.filter(Publisher.all(), this.getClass()));
-    }
-
-    private void buildPublishers( StaplerRequest req, JSONObject json, List<Descriptor<Publisher>> descriptors) throws FormException{
-        for (Descriptor<Publisher> d : descriptors) {
-            String name = d.getJsonSafeClassName();
-            BaseProjectProperty<Publisher> baseProperty = getBaseProjectProperty(name);
-            Publisher publisher = null;
-            if (json.has(name)) {
-                publisher = d.newInstance(req, json.getJSONObject(name));
-            }
-            baseProperty.setValue(publisher);
-        }
-    }
-
-    @Override
-    protected List<Action> createTransientActions() {
-        List<Action> r = super.createTransientActions();
-
-        for (BuildStep step : getBuildersList())
-            r.addAll(step.getProjectActions(this));
-        for (BuildStep step : getPublishersList())
-            r.addAll(step.getProjectActions(this));
-        for (BuildWrapper step : getBuildWrappers().values())
-            r.addAll(step.getProjectActions(this));
-        for (Trigger trigger : getTriggers().values())
-            r.addAll(trigger.getProjectActions());
-
-        return r;
     }
 
     /**
